requirements.txt
fastapi[standard]==0.113.0
sqlalchemy==2.0.41
pydantic-settings==2.9.1
cryptography==45.0.3
bcrypt
PyJWT
passlib